MySnake refactor :
Created new class for MySnake and moved code from MyFrame class to its own class.

After doing the eaten method in Food class did not work as it could not access the changeLength method from MyFrame anymore.

I corrected the path from MyFrame.mySnake.changeLength to MySnake.changeLength however, it could not access the method in the new class MySnake as it was not static.

I made the changeLength method static however the variable length also needed to be static to be used within this method.

Once both were changed to static the snake was able to eat food and grow in size.