SnakeObject refactor :
Created new class for SnakeObject and moved code from MyFrame class to its own class.

After doing this both the Food and MySnake class produced an error as they could not extend from the path MyFrame.SnakeObject.

To resolve this i changed the path from 'extends MyFrame.SnakeObject' to 'extends SnakeObject'.

This fixed the issue and the game ran correctly.