Scene Switch feat :
Previously implemented start button to load Play class and continue game from there however, this caused the current window to be closed and a new one opened.

To fix this I created a Main class to set the scene and load the StartScene.fxml file which included the start button. When pressed this start button loads another scene using the SceneController class.

This needs to be developed further to run the game in the second scene.