Playername fix :
Move original code provided to source package and created seperate controllers for StartScene and GameScene.
This enabled users to enter playername on the StartScene which is then displayedon the GameScene at the top left of the screen.

Next I need to use the provided source code to implement the working game version of the program.