Key pressed feat :
With the keyPressed method created in the GameSceneController class the program is able to regonise when a user presses WASD and output this information via command line.

The next step from this would be using this keyPressed method to move the snake in the game. Utilising the provided source code I will convert the swing approach to javafx.