Snake move feat :
Using SceneBuilder a circle object with the ID SnakeHeadCircle was added to the scene.
Through the keyPressed method in the GameSceneController class I used the setCenterX and Y method to increase or decrease the x and y of the circle object depending on the keys pressed.

This allows the circle object to move around the scene smoothly. I need to develop this further and put checks in place to ensure the snake object remains in the scene dimentions.