Snake head and tail move fix : When using the circle shape the snakeTail of the snake overlapped the snakeHead due to the edge of the circle object. To fix this issue the javafx rectangle object was used instead to create an even sided square for the snake.
This allowed the snakeTail to follow behind the snakeHead without overlapping.

For further development I would like to use a sprite to add different themes choices for the user however, at the moment the focus is on creating an outOfBounds method to ensure the snake object remains within the window screen.