EndScene fix: no thread error messages
The EndScene would previously be called from the switchToEndScene method in the GameSceneController however, though the same continuted in the IDE errors on the command line would occur indicating an issue with threads.

To fix this I had to specify in the switchToEndScene method "timeline.stop()". This fixes the issue as when the screen changes it stop timeline from checking every 0.3 if an action has occurred since that scene no longer exists.