Food generate feat: food object using random x y
foodGenerate method created in GameSceneController class to be called
when foodExists == false.

Within this method random numbers are generated for both x and y
coordinates that are later used to set the position of the foodObject.

The foodObject is set to use an image located in the images package.

After the object is created foodExists is set to true and foodObject is
added to the GameScene.