Eaten method feat : add snake tail when food eaten
Created foodEaten method to be called if foodExists variable is true.

This is done by getting the bounds in parent of both objects and
checking if they intersect. If the objects do intersect "FOOD EATEN" is
printed in command line and foodExists is set to false.

The addSnakeTail method is then called to add a tail onto the existing
snake body.