BodyHit Feat : ends game when snake hits itself
Created BodyHit method to check if the snake hits itself by comparing
the x and y coordinates of each headPoint in the linked list array. If
any are the same this means a hit has occurred.

When this happens, the method returns true that in turn calls the
SwitchToEndScene method to end the game.