Theme feat : snakeHead and snakeTail theme
Created Theme class that has the methods GenerateSnakeHeadImage and
GenerateSnakeTailImage. Using the themeNumber these methods set the
snakeHead or snakeTail image to be the corresponding image with the
users theme choice.

For example, if theme Red is chosen, the GenerateSnakeHeadImage method
is called with themeNumber passed. If themeNumber is 3 it sets the
snakeHeadImage to be head-red.png and returns this image.