Leaderboard feat : stores line to file
A leaderboard class was created to deal with the reading and writing of
scores to a textfile.

The WriteLeaderboardFile method in this class opens the leaderboard.txt
file and writes a new line then the phrase "WRITE TO FILE". The file is
then closed and a commandline printed to inform me that the task has
been completed.