Leaderboard fix : sorted scores in order
In the LeaderboardSceneController class the ScoreArraySort method is
used to sort the array.

Using a for loop the playerscoreList is looped through and sorted into
descending order with the playernameList doing the same for the
corresponding index.

The name and score labels in the leaderboard scene are then set using
the first 5 indexs of the score and name list to display the top 5
player scores to the user.