Theme fix : pacman theme added
The theme of pacman added as a theme choice options. This theme choice
causes the background to be set to pacman-background, the snakeHead to
be set to pacman-head and body to be set to pacman-body in the
initialize method.

Within the Obstacle class the theme choice is also affected, using a
switch case within the GenerateObstacle method to set the image to a
random pacman ghost.