Remove tail feat : obstacle hit removes tail
In the GameSceneController class the Snake.RemoveSnakeTail method is
called when the obstacle is hit to get the end tail rectangle in the
snakeBody. This is returned a rectangle object to be removed from the
 gameAnchorPane. The end index was also removed from the snakeBody
 linkedlist.

Before this an if statement had to be added to run the SwitchToEndScene
method if the playerScore is less than 0.