End scene fix : redesigned the end scene
Using scenebuilder I redesigned the endscene to include a label for the
playerscore and playername. In the EndSceneController the initialize
method was used to set these labels to the current players name and
score.

Buttons were added to allow the user to navigate to the start scene,
leaderboard scene or exit the game.