levelscene feat : scene to be shown between levels
I created a levelscene fxml file which takes the playerScore and
levelNumber and displays this to the player.

This scene is switched to when the user reaches a certain score. The
levelNumber is then increased in the LevelSceneController class. This
speeds up the movement of the snake, increasing the games difficultly as
 the player progress through the levels.