README chore : adding how to use project section
Within the README.md file I have added a 'How to use the project'
section explaining the controls for the game i.e. how to move the snake
and also what the food and obstacle objects do.

With food hits increasing score and length and obstacle hits decreasing
score and length.