level feat : obstacles + food respawn
Between levels instead of switching scenes the LevelPopup method is
called. This method pauses the timeline and shows the level pane,
displaying to the player the level they completed and their score.

The pane is the set to be invisible and the food and obstacle respawn
rate increased. This generates food and obstacle objects faster
increasing the games difficulty as the player progresses.